Fix some narrowing conversion warnings
git-svn-id: https://svn.isis.vanderbilt.edu/GMESRC/trunk@2709 b35830b6-564f-0410-ac5f-b387e50cb686

+++ GME/Common/CommonStl.h
@@ -79,7 +79,7 @@ inline bstr_const_iterator end(BSTR p) { return reinterpret_cast<wchar_t*>(p) ? 
 inline void CopyTo(bstr_const_iterator i, bstr_const_iterator e, CComBstrObj &b)
 {
 	ASSERT( i <= e );
-	BSTR tmpb = SysAllocStringLen(i, e - i);
+	BSTR tmpb = SysAllocStringLen(i, (UINT)(e - i));
 	if (tmpb == NULL)
 		COMTHROW(E_OUTOFMEMORY);
